split out memorypak and about screen to seperate modules
